+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add comment to README about rules being disabled by default.
+- Improve error output for invalid and missing configuration files.
 
 ## 0.0.12 (2023-05-30)